% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_loopy_results.R
\name{check_job_status}
\alias{check_job_status}
\title{Check job status}
\usage{
check_job_status(job_id, logging = TRUE)
}
\arguments{
\item{job_id}{string Loopy generated id for preparing or caching data requests.}

\item{logging}{boolean If set to TRUE will write output to logfile.}
}
\value{
numeric/string HTTP status code or job status progress.
}
\description{
Checks if a Loopy API request is in progress, finished, or failed.
Function needs quite a bit of checking.
}
