% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_loopy_results.R
\name{download_tracking_data}
\alias{download_tracking_data}
\title{Download Tracking Data}
\usage{
download_tracking_data(asset_id, group_key)
}
\arguments{
\item{asset_id}{string The asset_id that Loopy creates for each result.}

\item{group_key}{string The group_key that Loopy creates for each result.}
}
\value{
data.frame Tracking data.
}
\description{
Downloads and returns tracking data as a data frame.
}
