% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopy_general_calls.R
\name{list_videos}
\alias{list_videos}
\title{List Videos}
\usage{
list_videos(video_type = c("uploaded", "results"), verbose = "0")
}
\arguments{
\item{video_type}{string Videos that were either uploaded to Loopy ("uploaded") or produced by Loopy's video processing ("results").}

\item{verbose}{string A URL parameter which permits user access to all the files they have permission to access ('1') or to their files ('0'); default is '0'.}
}
\value{
list of Loopy videos with metadata.
}
\description{
Lists videos hosted on Loopy, list can either be of videos uploaded to Loopy or those that were produced via a Loopy video processing operation.
}
