% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopy_api.R
\name{loopy_api}
\alias{loopy_api}
\title{Call Loopy API}
\usage{
loopy_api(
  url,
  verbose = "0",
  video_id = NULL,
  collection_id = NULL,
  group_key = NULL,
  asset_id = NULL,
  format = NULL
)
}
\arguments{
\item{url}{string Loopy URL file path (origin should be set with set_loopy_user)}

\item{verbose}{string A URL parameter which permits user access to all the files they have permission to access ('1') or to their files ('0'); default is '0'.}

\item{video_id}{string A URL parameter which limits return to only those with the video id.}

\item{collection_id}{string A URL parameter which limits the return to only those from the specified collection.}
}
\value{
list A list including parse JSON (or HTML) data, the URL used in the API request, and the status code of the request.
}
\description{
Wrapper for httr::GET that handles Loopy API requests.
}
