% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopy_user.R
\name{set_loopy_user}
\alias{set_loopy_user}
\title{Set Loopy user}
\usage{
set_loopy_user(api_key, loopy_origin, renviron_path = Sys.getenv("HOME"))
}
\arguments{
\item{api_key}{string Users API key.}

\item{loopy_origin}{string Loopy URL origin (protocol, subdomain, domain name and port, where applicable); Paths, querys, and parameters are specified in other functions.}

\item{renviron_path}{string Path where .Renviron is stored. Default is the user's home directory.}
}
\value{
Creates (if necessary) and appends user's Loopy API key and URL to user's .Renviron file.
}
\description{
Stores Loopy API key and URL in .Renviron, creates .Renviron if it doesn't exist.
}
